refactor YulStringGenerator into methods in yulAST according to PR comments
